Moved syncmap I/O into aeneas.syncmap subpackage
