+++ MANIFEST.in
@@ -7,6 +7,7 @@ recursive-include aeneas/cwave *
 recursive-include aeneas/extra *
 prune aeneas/extra/ctw_speect
 recursive-include aeneas/res *
+recursive-include aeneas/syncmap *
 recursive-include aeneas/tools/res *
 recursive-include aeneas/ttswrappers *
 include aeneas_check_setup.py
