+++ aeneas/syncmap/smfjson.py
@@ -1,0 +1,100 @@
+#!/usr/bin/env python
+# coding=utf-8
+
+# aeneas is a Python/C library and a set of tools
+# to automagically synchronize audio and text (aka forced alignment)
+#
+# Copyright (C) 2012-2013, Alberto Pettarin (www.albertopettarin.it)
+# Copyright (C) 2013-2015, ReadBeyond Srl   (www.readbeyond.it)
+# Copyright (C) 2015-2016, Alberto Pettarin (www.albertopettarin.it)
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU Affero General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+"""
+TBW
+"""
+
+from __future__ import absolute_import
+from __future__ import print_function
+import json
+
+from aeneas.syncmap.smfbase import SyncMapFormatBase
+import aeneas.globalfunctions as gf
+
+
+class SyncMapFormatJSON(SyncMapFormatBase):
+
+    TAG = u"SyncMapFormatJSON"
+
+    DEFAULT = "json"
+    """
+    JSON::
+
+        {
+         "fragments": [
+          {
+           "id": "f001",
+           "language": "en",
+           "begin": 0.000,
+           "end": 1.234,
+           "children": [],
+           "lines": [
+            "First fragment text"
+           ]
+          },
+          {
+           "id": "f002",
+           "language": "en",
+           "begin": 1.234,
+           "end": 5.678,
+           "children": [],
+           "lines": [
+            "Second fragment text",
+            "Second line of second fragment"
+           ]
+          },
+          {
+           "id": "f003",
+           "language": "en",
+           "begin": 5.678,
+           "end": 7.890,
+           "children": [],
+           "lines": [
+            "Third fragment text",
+            "Second line of third fragment"
+           ]
+          }
+         ]
+        }
+
+    * Multiple levels: yes (output only)
+    * Multiple lines: yes
+
+    .. versionadded:: 1.2.0
+    """
+
+    def parse(self, input_text, syncmap):
+        contents_dict = json.loads(input_text)
+        for fragment in contents_dict["fragments"]:
+            self._add_fragment(
+                syncmap=syncmap,
+                identifier=fragment["id"],
+                language=fragment["language"],
+                lines=fragment["lines"],
+                begin=gf.time_from_ssmmm(fragment["begin"]),
+                end=gf.time_from_ssmmm(fragment["end"])
+            )
+
+    def format(self, syncmap):
+        return syncmap.json_string
