+++ aeneas/tests/res/syncmaps/sonnet001_m.eaf
@@ -1,5 +1,5 @@
 <?xml version='1.0' encoding='UTF-8'?>
-<ANNOTATION_DOCUMENT xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="http://www.mpi.nl/tools/elan/EAFv2.8.xsd" AUTHOR="aeneas" DATE="2016-10-18T17:25:07+00:00" FORMAT="2.8" VERSION="2.8">
+<ANNOTATION_DOCUMENT xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="http://www.mpi.nl/tools/elan/EAFv2.8.xsd" AUTHOR="aeneas" DATE="2016-10-19T16:21:38+00:00" FORMAT="2.8" VERSION="2.8">
   <HEADER MEDIA_FILE="" TIME_UNITS="milliseconds"/>
   <TIME_ORDER>
     <TIME_SLOT TIME_SLOT_ID="ts000001b" TIME_VALUE="0"/>
