+++ aeneas/tests/res/syncmaps/sonnet001_mu.sbv
@@ -1,4 +1,5 @@
 [SUBTITLE]
+
 00:00:00.000,00:00:02.680
 I
